Fixed how relative source paths and id prefixes are computed.
The relative paths used as source paths in export files and for
prefixing ids in content files, were previously always computed based on
the current working directory. This has changed, such that the paths are
now correctly computed based on the base path inferred from globs and
the 'base' option specified for the Gulp 'src' method.

To handle those cases where that is not the desired behavior, the option
'baseFilePath' has been added to the export and import config
interfaces. To preserve the old behavior, i.e. paths relative to cwd,
simply set that option to "." in both the import and export config.