Refactoring, linting and bugfix.
Refactored the plugin so it can be used without gulp. The intention is
to eventually move the core code into a separate package, which may then
be used by plugins for different build systems, such as gulp and
webpack. Also added linting and fixed a bug causing missing translations
for files that have more than one root element.