Replaced the abstract 'File' class with an interface.
This makes it easier to use the plugin without Gulp, as the tasks can
now process files specified as simple object literals conforming to the
new 'IFile' interface.

The `base` property on the 'IFile' interface has also been made
optional, as e.g. Webpack does not provide a base path for each file,
meaning it has to be configured in the task options instead.